+++ src/constants.rs
@@ -12,7 +12,7 @@ pub const USER_PREFIX: &str = "user";
 pub const ERROR_FILES_DIR: &str = ".error_files";
 
 pub const METAPLEX_PROGRAM_ID: &str = "metaqbxxUerdq28cj1RbAWkYQm3ybzjb6a8bt518x1s";
-pub const CANDY_MACHINE_PROGRAM_ID: &str = "cndyAnrLdpjq1Ssp1z8xxDsB8dxe7u4HL5Nxi2K5WXZ";
+pub const TARS_PROGRAM_ID: &str = "cndyAnrLdpjq1Ssp1z8xxDsB8dxe7u4HL5Nxi2K5WXZ";
 
 pub const PUBLIC_RPC_URLS: &[&str] = &[
     "https://api.devnet.solana.com",